requirements.txt
aiogram>=3.0
pyyaml>=6.0